Address review: int64 fallback dtype for maxlen overflow safety, add uint8 lengths test coverage

+++ tensorflow/python/ops/array_ops.py
@@ -4177,7 +4177,7 @@ def sequence_mask(lengths, maxlen=None, dtype=dtypes.bool, name=None):
     # Because of broadcasting on both arguments this comparison results
     # in a matrix of size (len(lengths), maxlen)
     if not maxlen.dtype.is_integer:
-      maxlen = gen_math_ops.cast(maxlen, dtypes.int32)
+      maxlen = gen_math_ops.cast(maxlen, dtypes.int64)
     maxlen_dtype = maxlen.dtype
     row_vector = gen_math_ops._range(
         constant(0, maxlen_dtype), maxlen, constant(1, maxlen_dtype))
